Reorganized Folders and Added Sprite Sheet
I've completely reorganized our repo since it was a mess before, please
make sure to submit work into this system appropriately in the future
rather than just dumping it in.

Created and added sprite sheet for the moving parts of our game. I
realize now that it is not viable to put EVERY asset in a big sprite
sheet because the max resolution for a sprite sheet is like 2000x2000
px.

I've also updated many of Lucas's assets to make them look "good" and
labeled them with DQnew at the end of the file name.